Registration now has registrationId - not ticketId Same for changes. Requests will have to change to send "regid" rather than "ticketid" Also actions that referred to ticket now have to refer to reg (see mapping below)
A lot of the jsp and js refers to ticket when it should refer to reg[istration] - changed some but not all.

Associated with the registration is a collection of tickets - each has a unique uuid - could be used for verification etc.

The type can remain - limited now to "reg" or "hold"
Should never send the type as a request parameter - it's derived from the action

Users are allocated what's available - they can ask for as many as they want up to the event limit.
At the point they ask we won't refuse

They can be allocated less than the request - means on the waitq.

Two fields
  numTickets - how many they have
  ticketsRequested - how many they want

Added a waitqDate to order the waitq

Removed session manager isWaiting and isRegistered methods. They should be on the registration anyway

Need some new event parameters to control some of this behavior. max waitq length for one

Update isn't checking the deadlines. If the user is trying to increase the tickets should disallow if the deadline has passed?